factor out common lens code
