+++ lens/util/store.go
@@ -1,4 +1,4 @@
-package carrepo
+package util
 
 import (
 	"context"
